@@ -19,9 +19,9 @@
     "moment": "^2.22.2",
     "notistack": "^0.3.9",
     "prop-types": "^15.5.10",
-    "react": "^16.6.0",
+    "react": "^16.8.0-alpha.0",
     "react-color": "^2.14.1",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.8.0-alpha.0",
     "react-redux": "^5.0.7",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
